requirements.txt
streamlit
crewai
faiss-cpu
python-dotenv
docker
numpy